finished backend for commitments

+++ backend/src/repositories/commitments.repository.js
@@ -1,4 +1,6 @@
+/* eslint-disable max-len */
 import { db } from '../data/connection';
+import { challengeRepo } from './challenge.repository';
 
 export const commitmentsRepo = {
   async getCommitments() {
@@ -20,9 +22,11 @@ export const commitmentsRepo = {
       startDate,
       endDate,
     } = commitment;
-    const sqlQuery = 'INSERT INTO commitments (name, start_date, end_date, user_id) VALUES (?,?,?,?)';
+    const challengeQuery = await challengeRepo.getChallenge();
+    const challengeId = challengeQuery.id;
+    const sqlQuery = 'INSERT INTO commitments (name, start_date, end_date, user_id, challenge_id) VALUES (?,?,?,?,?)';
     try {
-      return await db.query(sqlQuery, [name, startDate, endDate, userId]);
+      return await db.query(sqlQuery, [name, startDate, endDate, userId, challengeId]);
     } catch (err) {
       throw {
         status: 500,
@@ -33,7 +37,7 @@ export const commitmentsRepo = {
   async getCommitment(id) {
     const sqlQuery = 'SELECT * FROM commitments WHERE id = ?';
     try {
-      const queryData = db.query(sqlQuery, id);
+      const queryData = await db.query(sqlQuery, id);
       return queryData.results[0];
     } catch (err) {
       throw {
@@ -42,10 +46,10 @@ export const commitmentsRepo = {
       };
     }
   },
-  async removeCommitment(id) {
-    const sqlQuery = 'DELETE FROM commitments WHERE id = ?';
+  async removeCommitment(id, userId) {
+    const sqlQuery = 'DELETE FROM commitments WHERE id = ? AND user_id = ?';
     try {
-      return await db.query(sqlQuery, id);
+      return await db.query(sqlQuery, [id, userId]);
     } catch (err) {
       throw {
         status: 500,
@@ -57,6 +61,30 @@ export const commitmentsRepo = {
     const sqlQuery = 'DELETE FROM commitments WHERE name = ? AND user_id = ?';
     try {
       return await db.query(sqlQuery, [commitmentName, userId]);
+    } catch (err) {
+      throw {
+        status: 500,
+        message: err.sqlMessage,
+      };
+    }
+  },
+  async updateCommitment(commitment) {
+    const {
+      startDate,
+      endDate,
+      name,
+      isDone,
+      id,
+    } = commitment;
+    const sqlQuery = 'UPDATE commitments SET name = ?, start_date = ?, end_date = ?, is_done = ? WHERE id = ?';
+    try {
+      return await db.query(sqlQuery, [
+        name,
+        new Date(startDate).toISOString().slice(0, 19).replace('T', ' '),
+        new Date(endDate).toISOString().slice(0, 19).replace('T', ' '),
+        isDone,
+        id,
+      ]);
     } catch (err) {
       throw {
         status: 500,
